requirements.txt
fastapi
uvicorn
torch
torchvision
torchaudio
transformers
pillow